requirements.txt
PyYAML
gevent
Routes
WebOb
